Separando View del navegador
