+++ src/Views/ViewLookForWorker.js
@@ -1,7 +1,12 @@
+import { Container } from "reactstrap";
 
 
 export const ViewLookForWorker = () => {
     return (
-        <h1>Aca mostraria esta el formulario de busqueda de Servicos<a href="/findService/">Buscar</a></h1>
+        <Container className="abs-center">
+            
+            <a href="/findService/">Buscar</a>
+            Aca mostraria esta el formulario de busqueda de Servicos
+        </Container>
         );
     }
